Categorised the scripts into different folders
Previously, all the scripts started with the prefix 'sai_' and were all in the same folder. In this commit, I have removed the prefix 'sai_' to make it more formal and also grouped the scripts into different folder categories for better convenience. Going ahead, every new script will be put in the respective folders or new folders.
