+++ sai_scripts/function_utilities/read_obd_velocity_script.py
@@ -8,6 +8,27 @@ import numpy as np
 from scipy import interpolate
 import time
 
+def convert24(hrs_mins_secs,am_pm): 
+      
+    # Checking if last two elements of time 
+    # is AM and first two elements are 12 
+    if am_pm == "AM" and hrs_mins_secs[:2] == "12": 
+        return "00" + hrs_mins_secs[2::] 
+          
+    # remove the AM     
+    elif am_pm == "AM": 
+        return hrs_mins_secs
+      
+    # Checking if last two elements of time 
+    # is PM and first two elements are 12    
+    elif am_pm == "PM" and hrs_mins_secs[:2] == "12": 
+        return hrs_mins_secs 
+          
+    else: 
+          
+        # add 12 to hours and remove PM 
+        return str(int(hrs_mins_secs[:2]) + 12) + hrs_mins_secs[2::]
+
 
 fpath = 'D:\\Steradian\\Radar_Data\\recorded_data\\driveData_withinArcade_withoutLid_07.12.2018\\'
 flag_obd_data_available = 1
@@ -70,26 +91,7 @@ else:
     obd_vel_kmph = 0
 
 
-def convert24(hrs_mins_secs,am_pm): 
-      
-    # Checking if last two elements of time 
-    # is AM and first two elements are 12 
-    if am_pm == "AM" and hrs_mins_secs[:2] == "12": 
-        return "00" + hrs_mins_secs[2::] 
-          
-    # remove the AM     
-    elif am_pm == "AM": 
-        return hrs_mins_secs
-      
-    # Checking if last two elements of time 
-    # is PM and first two elements are 12    
-    elif am_pm == "PM" and hrs_mins_secs[:2] == "12": 
-        return hrs_mins_secs 
-          
-    else: 
-          
-        # add 12 to hours and remove PM 
-        return str(int(hrs_mins_secs[:2]) + 12) + hrs_mins_secs[2::]
+
