+++ sai_scripts/function_utilities/writeToTextFile.py
