+++ sai_scripts/non_stationary_signals/chirp_plus_sinosoid_iterference.py
@@ -47,17 +47,17 @@ plt.ylabel('Freq (GHz)')
 plt.grid(True)
 
 plt.figure(2)
-plt.subplot(331)
+plt.subplot(131)
 plt.title('FFT of chirp signal')
 plt.plot(freq_grid/1e6, 20*np.log10(np.fft.fftshift(np.abs(chirp_signal_fft))))
 plt.xlabel('Freq(MHz)')
 plt.grid(True)
-plt.subplot(332)
+plt.subplot(132)
 plt.title('FFT of sinusoid signal')
 plt.plot(freq_grid/1e6, 20*np.log10(np.fft.fftshift(np.abs(sinusoid_signal_fft))))
 plt.xlabel('Freq(MHz)')
 plt.grid(True)
-plt.subplot(333)
+plt.subplot(133)
 plt.title('FFT of interference corrupted signal')
 plt.plot(freq_grid/1e6, 20*np.log10(np.fft.fftshift(np.abs(corrupted_signal_fft))))
 plt.xlabel('Freq(MHz)')
@@ -65,13 +65,13 @@ plt.grid(True)
 
 
 plt.figure(3)
-plt.subplot(221)
+plt.subplot(121)
 plt.title('FFT using first half samples')
 plt.plot(freq_grid/1e6, 20*np.log10(np.fft.fftshift(np.abs(corrupted_signal_firsthalf_fft))))
 plt.xlabel('Freq(MHz)')
 plt.grid(True)
 
-plt.subplot(222)
+plt.subplot(122)
 plt.title('FFT using second half samples')
 plt.plot(freq_grid/1e6, 20*np.log10(np.fft.fftshift(np.abs(corrupted_signal_secondhalf_fft))))
 plt.xlabel('Freq(MHz)')
