+++ sai_scripts/radar_modeling/angle_resol_check.py
