+++ sai_scripts/radar_modeling/coherent_vs_noncoherent_avg.py
@@ -36,8 +36,8 @@ coh_avg = np.fft.fft(signal_range_fft,axis=1)[:,0]/num_samp_dopp
 est_noise_power_range_fft = 10*np.log10((np.sum(np.abs(signal_range_fft[:,0])**2) - np.amax(np.abs(signal_range_fft[:,0]))**2))
 est_noise_power_non_coh_avg = 10*np.log10((np.sum(non_coh_avg**2) - np.amax(non_coh_avg)**2))
 est_noise_power_coh_avg = 10*np.log10((np.sum(np.abs(coh_avg)**2) - np.amax(np.abs(coh_avg))**2))
-print('\n\n')
-print('\nTrue Noise Power:', 10*np.log10(noise_power), '\nNoise Power estimated from single chirp: ', est_noise_power_range_fft, '\nNoise Power non-coh Avg.: ', est_noise_power_non_coh_avg, '\nNoise Power coh Avg.:', est_noise_power_coh_avg)
+print('\n')
+print('\nTrue Noise Power: {0:.2f} dB \nNoise Power estimated from single chirp: {1:.2f} dB \nNoise Power non-coh Avg.: {2:.2f} dB \nNoise Power coh Avg.: {3:.2f} dB'.format(10*np.log10(noise_power), est_noise_power_range_fft,  est_noise_power_non_coh_avg, est_noise_power_coh_avg))
 range_grid = np.arange(-np.pi,np.pi,2*np.pi/num_samp_range)
 plt.figure(1)
 plt.plot(range_grid,20*np.log10(np.fft.fftshift(np.abs(coh_avg))),alpha=0.5)
