+++ sai_scripts/radar_modeling/radar_model_cdm_testbench.py
